Update test coverage, rename file
