+++ pyvista/examples/_example_loader.py
